Remove deprecated connection classes
