+++ phpstan.neon.dist
@@ -19,7 +19,7 @@ parameters:
         # changing these would be a BC break, to be done in next major
         - "~^Casting to bool something that's already bool.~"
         - '~^Property Doctrine\\DBAL\\Schema\\Schema::\$_schemaConfig \(Doctrine\\DBAL\\Schema\\SchemaConfig\) does not accept default value of type false\.\z~'
-        - '~^Return type \(int\|false\) of method Doctrine\\DBAL\\Driver\\OCI8\\OCI8Connection\:\:lastInsertId\(\) should be compatible with return type \(string\) of method Doctrine\\DBAL\\Driver\\Connection::lastInsertId\(\)~'
+        - '~^Return type \(int\|false\) of method Doctrine\\DBAL\\Driver\\OCI8\\Connection\:\:lastInsertId\(\) should be compatible with return type \(string\) of method Doctrine\\DBAL\\Driver\\Connection::lastInsertId\(\)~'
 
         # https://github.com/phpstan/phpstan/issues/2857
         # TODO: remove in 4.0.0
@@ -48,12 +48,12 @@ parameters:
         # Requires a release of https://github.com/JetBrains/phpstorm-stubs/pull/553
         -
             message: '~^Call to function assert\(\) with true will always evaluate to true\.$~'
-            path: %currentWorkingDirectory%/src/Driver/PDOConnection.php
+            path: %currentWorkingDirectory%/src/Driver/PDO/Connection.php
 
         # Requires a release of https://github.com/JetBrains/phpstorm-stubs/pull/553
         -
             message: '~^Strict comparison using !== between int and false will always evaluate to true\.$~'
-            path: %currentWorkingDirectory%/src/Driver/PDOConnection.php
+            path: %currentWorkingDirectory%/src/Driver/PDO/Connection.php
 
         # False Positive
         - '~Strict comparison using === between 1 and 2 will always evaluate to false~'
